infer_realtime/requirements.txt
absl-py==1.2.0
altair==4.2.0
amplitude-tracker==0.0.7
anyio==3.6.1
astunparse==1.6.3
attrs==21.4.0
backoff==1.10.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bidict==0.22.0
black==22.10.0
blinker==1.5
boto3==1.24.92
botocore==1.27.92
cachetools==5.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.1.3
cloudpickle==2.2.0
commonmark==0.9.1
contourpy==1.0.5
coolname==1.1.0
cramjam==2.5.0
cycler==0.11.0
databricks-cli==0.17.3
decorator==5.1.1
distlib==0.3.3
entrypoints==0.4
fastapi==0.78.0
fastparquet==0.7.2
filelock==3.0.12
flatbuffers==22.9.24
fonttools==4.37.4
fsspec==2022.8.2
gast==0.4.0
gitdb==4.0.9
gitignorefile==1.1.2
GitPython==3.1.28
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h11==0.14.0
h5py==3.7.0
idna==3.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==3.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.18.0
matplotlib==3.6.1
mistune==0.8.4
mlfoundry==0.4.10
monotonic==1.6
mypy-extensions==0.4.3
numpy==1.21.1
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pathspec==0.10.1
Pillow==9.2.0
pipenv==2021.5.29
platformdirs==2.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.2
puremagic==1.14
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pydeck==0.8.0b3
Pygments==2.12.0
PyJWT==2.4.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.3
python-multipart==0.0.5
python-socketio==5.7.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
questionary==1.10.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rich-click==1.5.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.2
semver==2.13.0
servicefoundry==0.2.19
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.19.1
streamlit==1.13.0
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tfy-mlflow-client==0.0.17
threadpoolctl==3.1.0
thrift==0.16.0
tinynetrc==1.3.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing-extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.6
uvicorn==0.18.3
validators==0.20.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
websocket-client==1.3.3
Werkzeug==2.2.2
whylabs-client==0.3.0
whylabs-datasketches==2.2.0b1
whylogs==0.6.30
wrapt==1.14.1
zipp==3.8.1
